requirements.txt
transformers==4.41.2
huggingface_hub==0.23.4
safetensors
pillow
numpy
runpod